Nuke `ReferenceStub`
