+++ analyzer-gen2/src/main/kotlin/com/github/cubuspl42/sigmaLang/shell/terms/ReferenceTerm.kt
@@ -1,13 +1,13 @@
 package com.github.cubuspl42.sigmaLang.shell.terms
 
 import com.github.cubuspl42.sigmaLang.analyzer.parser.antlr.SigmaParser
-import com.github.cubuspl42.sigmaLang.core.ShadowExpression
+import com.github.cubuspl42.sigmaLang.core.ExpressionBuilder
 import com.github.cubuspl42.sigmaLang.core.expressions.Expression
 import com.github.cubuspl42.sigmaLang.core.values.Identifier
 import com.github.cubuspl42.sigmaLang.core.values.UnorderedTupleValue
 import com.github.cubuspl42.sigmaLang.core.values.Value
+import com.github.cubuspl42.sigmaLang.shell.FormationContext
 import com.github.cubuspl42.sigmaLang.shell.stubs.ExpressionStub
-import com.github.cubuspl42.sigmaLang.shell.stubs.ReferenceStub
 
 data class ReferenceTerm(
     val referredName: IdentifierTerm,
@@ -22,9 +22,18 @@ data class ReferenceTerm(
         override fun extract(parser: SigmaParser): SigmaParser.ReferenceContext = parser.reference()
     }
 
-    override fun transmute(): ExpressionStub<Expression> = ReferenceStub(
-        referredName = referredName.transmute(),
-    )
+    override fun transmute(): ExpressionStub<Expression> = object : ExpressionStub<Expression>() {
+        override fun transform(
+            context: FormationContext,
+        ): ExpressionBuilder<Expression> {
+            val scope = context.scope
+
+            val referredExpression = scope.resolveName(referredName = referredName.toIdentifier())
+                ?: throw IllegalStateException("Unresolved reference: $referredName")
+
+            return ExpressionBuilder.pure(referredExpression)
+        }
+    }
 
     override fun wrap(): Value = UnorderedTupleValue(
         valueByKey = mapOf(
